+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+#### Fixed
+
+- Name variable & code list selection is now mandatory for pairing question type
+
 ## [3.3.0](https://github.com/InseeFr/Pogues/releases/tag/3.3.0) - 2026-06-11
 
 ### Added